attack speed fix

+++ app/src/main/java/org/o7planning/android2dgame/OrcAI.java
@@ -83,7 +83,7 @@ public class OrcAI implements CharacterAI {
                     character.hurtBox.x + character.hurtBox.width > other.hitBox.x &&
                     character.hurtBox.y < other.hitBox.y + other.hitBox.height &&
                     character.hurtBox.y + character.hurtBox.height > other.hitBox.y){
-                other.reduceHitPointsBy((double) character.attackDamage);
+                other.reduceHitPointsBy(character.attackDamage);
                 character.isAttacking = true;
             }
         }
